+++ src/process/validation/rules/__tests__/validateCustom.test.ts
@@ -4,7 +4,7 @@ import { FieldError } from 'error';
 import { TextFieldComponent } from 'types';
 import { simpleTextField } from './fixtures/components';
 import { validateCustom } from '../validateCustom';
-import { generateProcessContext } from './fixtures/util';
+import { generateProcessorContext } from './fixtures/util';
 
 it('A simple custom validation will correctly be interpolated', async () => {
     const component: TextFieldComponent = {
@@ -16,7 +16,7 @@ it('A simple custom validation will correctly be interpolated', async () => {
     const data = {
         component: 'any thing',
     }
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateCustom(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result && result.errorKeyOrMessage).to.equal('Invalid entry');
@@ -32,7 +32,23 @@ it('A custom validation that includes data will correctly be interpolated', asyn
     const data = {
         simpleComponent: 'any thing',
     }
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateCustom(context);
     expect(result).to.equal(null);
+});
+
+it('A custom validation of empty component data will still validate', async () => {
+    const component: TextFieldComponent = {
+        ...simpleTextField,
+        validate: {
+            custom: 'valid = data.simpleComponent === "any thing" ? true : "Invalid entry"',
+        },
+    };
+    const data = {
+        simpleComponent: '',
+    };
+    const context = generateProcessorContext(component, data);
+    const result = await validateCustom(context);
+    expect(result).to.be.instanceOf(FieldError);
+    expect(result && result.errorKeyOrMessage).to.equal('Invalid entry');
 });