+++ src/process/validation/rules/__tests__/validateEmail.test.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 import { FieldError } from 'error';
 import { simpleEmailField } from './fixtures/components';
-import { generateProcessContext } from './fixtures/util';
+import { generateProcessorContext } from './fixtures/util';
 import { validateEmail } from '../validateEmail';
 
 it('Validating a valid email will return null', async () => {
@@ -10,7 +10,7 @@ it('Validating a valid email will return null', async () => {
     const data = {
         component: 'sales@form.io',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateEmail(context);
     expect(result).to.equal(null);
 });
@@ -20,7 +20,7 @@ it('Validating an invalid email will return a FieldError', async () => {
     const data = {
         component: 'salesatform.io',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateEmail(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.contain('invalid_email');