+++ src/process/validation/rules/__tests__/validateMaximumDay.test.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 import { FieldError } from 'error';
 import { simpleDayField, simpleTextField } from './fixtures/components';
-import { generateProcessContext } from './fixtures/util';
+import { generateProcessorContext } from './fixtures/util';
 import { validateMaximumDay } from '../validateMaximumDay';
 
 it('Validating a non-day component will return null', async () => {
@@ -10,7 +10,7 @@ it('Validating a non-day component will return null', async () => {
     const data = {
         component: 'Hello, world!',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMaximumDay(context);
     expect(result).to.equal(null);
 });
@@ -20,7 +20,7 @@ it('Validating a day component with a day after the maximum day will return a Fi
     const data = {
         component: '04/02/2023',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMaximumDay(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('maxDay');
@@ -31,7 +31,7 @@ it('Validating a day component with a day before the maximum day will return nul
     const data = {
         component: '03/23/2023',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMaximumDay(context);
     expect(result).to.equal(null);
 });
@@ -41,7 +41,7 @@ it('Validating a day component with a day after the maximum day will return a Fi
     const data = {
         component: '04/02/2023',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMaximumDay(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('maxDay');
@@ -52,7 +52,7 @@ it('Validating a day-first day component with a day after the maximum day will r
     const data = {
         component: '02/04/2023',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMaximumDay(context);
     expect(result).to.be.instanceOf(FieldError);
     expect(result?.errorKeyOrMessage).to.equal('maxDay');
@@ -63,7 +63,7 @@ it('Validating a day-first day component with a day before the maximum day will 
     const data = {
         component: '23/03/2023',
     };
-    const context = generateProcessContext(component, data);
+    const context = generateProcessorContext(component, data);
     const result = await validateMaximumDay(context);
     expect(result).to.equal(null);
 });