+++ src/process/validation/rules/databaseRules.ts
@@ -1,7 +1,9 @@
 import { ValidationRuleInfo } from "types";
 import { validateUniqueInfo } from "./validateUnique";
+import { validateCaptchaInfo } from "./validateCaptcha";
 
 // These are the validations that require a database connection.
-export const DatabaseRules: ValidationRuleInfo[] = [
+export const databaseRules: ValidationRuleInfo[] = [
     validateUniqueInfo,
+    validateCaptchaInfo
 ];