+++ src/process/validation/rules/validateTime.ts
@@ -4,6 +4,9 @@ import { isEmpty } from "../util";
 import { FieldError, ValidatorError } from 'error';
 import { dayjs } from 'utils/date';
 import { ProcessorInfo } from "types/process/ProcessorInfo";
+import customParsers from 'dayjs/plugin/customParseFormat';
+
+dayjs.extend(customParsers);
 
 const isValidatableTimeComponent = (comp: any): comp is TimeComponent => {
     return comp && comp.type === 'time';
@@ -18,14 +21,17 @@ export const shouldValidate = (context: ValidationContext) => {
 };
 
 export const validateTimeSync: RuleFnSync = (context: ValidationContext) => {
-    const { component, value } = context;
+    const { component, value, config } = context;
     if (!shouldValidate(context)) {
         return null;
     }
     try {
         if (!value || isEmpty(component, value)) return null;
-        const format = (component as TimeComponent).format || 'HH:mm';
-        const isValid = dayjs(String(value), format).isValid();
+        // Server side evaluations of validity should use the "dataFormat" vs the "format" which is used on the client.
+        const format = config?.server ?
+            ((component as TimeComponent).dataFormat || 'HH:mm:ss') :
+            ((component as TimeComponent).format || 'HH:mm');
+        const isValid = dayjs(String(value), format, true).isValid();
         return isValid ? null : new FieldError('time', context);
     }
     catch (err) {