+++ src/types/Component.ts
@@ -163,6 +163,17 @@ export type DataSourceComponent = BaseComponent & {
     };
 };
 
+export type DataTableComponent = EditGridComponent & {
+    fetch?: {
+        enableFetch: boolean;
+        dataSrc: 'resource' | 'url';
+        sort?: { defaultQuery?: string };
+        resource?: string;
+        headers?: { key: string; value: string }[];
+        components?: { key: string; path: string }[] | Component[];
+    }
+}
+
 export type DateTimeComponent = BaseComponent & {
     format?: string;
     useLocaleSettings?: boolean;
@@ -483,7 +494,7 @@ export type TextAreaComponent = TextFieldComponent & {
 };
 
 export type TimeComponent = TextFieldComponent & {
-    format: string;
+    format?: string;
     dataFormat: string;
 };